Fix stream issues, work in progress
The main stream issues seem to be fixed now.
I'm avoiding use of coroutines altogether in favor of a kind of
cheesy callback construct for streams.  It seems to work.
It's basically a sequins but with resetting and memory.
I thought the memory would be cool because I see we "cache"
a lot of stream values, and it seems nice to have that built in.

The table stream (which is basically just a sequins) is broken.
It has the reset/cache stuff which I think will be nice. But
it always returns nil right now which is obviously suboptimal.