Add loop increment, last nil check
Add increment option for looping to do "broken modulo", for i.e.
adding a mutually prime increment to a stream. Good for melodies.

Also add a nil check to the "last" method for a stream, so that
when it hasn't been initialized by calling "next", it will still
obtain and return a value, providing a sensible default while
avoiding errors.